control-service: split DataJobTerminationStatusIT
Splitted DataJobTerminationStatusIT into DataJobExecutionIT and
DataJobTerminationStatusMonitoringIT.

Testing Done: integration tests
